Use jinja2 for the email template
